+++ tgbotapi.extensions.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/triggers_handling/CommandHandling.kt
@@ -12,6 +12,7 @@ suspend fun BehaviourContext.command(
     requireOnlyCommandInMessage: Boolean = true,
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     additionalFilter: CommonMessageFilter<TextContent>? = null,
+    performInParallel: Boolean = true,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, CommonMessage<TextContent>>
 ): Job = onText(
     includeFilterByChatInBehaviourSubContext,
@@ -27,6 +28,7 @@ suspend fun BehaviourContext.command(
             commandRegex.matches(it.asBotCommandTextSource() ?.command ?: return@any false)
         } && (additionalFilter ?.invoke(message) != false)
     },
+    performInParallel,
     scenarioReceiver
 )
 suspend fun BehaviourContext.command(
@@ -34,21 +36,24 @@ suspend fun BehaviourContext.command(
     requireOnlyCommandInMessage: Boolean = true,
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     additionalFilter: CommonMessageFilter<TextContent>? = null,
+    performInParallel: Boolean = true,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, CommonMessage<TextContent>>
-) = command(command.toRegex(), requireOnlyCommandInMessage, includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+) = command(command.toRegex(), requireOnlyCommandInMessage, includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)
 
 suspend inline fun BehaviourContext.onCommand(
     commandRegex: Regex,
     requireOnlyCommandInMessage: Boolean = true,
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     noinline additionalFilter: CommonMessageFilter<TextContent>? = null,
+    performInParallel: Boolean = true,
     noinline scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, CommonMessage<TextContent>>
-): Job = command(commandRegex, requireOnlyCommandInMessage, includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+): Job = command(commandRegex, requireOnlyCommandInMessage, includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)
 
 suspend inline fun BehaviourContext.onCommand(
     command: String,
     requireOnlyCommandInMessage: Boolean = true,
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     noinline additionalFilter: CommonMessageFilter<TextContent>? = null,
+    performInParallel: Boolean = true,
     noinline scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, CommonMessage<TextContent>>
-): Job = onCommand(command.toRegex(), requireOnlyCommandInMessage, includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+): Job = onCommand(command.toRegex(), requireOnlyCommandInMessage, includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)