+++ tgbotapi.extensions.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/triggers_handling/InlineQueryTriggers.kt
@@ -3,6 +3,7 @@ package dev.inmo.tgbotapi.extensions.behaviour_builder.triggers_handling
 import dev.inmo.micro_utils.coroutines.subscribeSafelyWithoutExceptions
 import dev.inmo.tgbotapi.extensions.behaviour_builder.*
 import dev.inmo.tgbotapi.extensions.behaviour_builder.expectations.expectFlow
+import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.optionallyWrapWithLaunch
 import dev.inmo.tgbotapi.extensions.utils.asInlineQueryUpdate
 import dev.inmo.tgbotapi.extensions.utils.extensions.sourceChat
 import dev.inmo.tgbotapi.types.InlineQueries.query.*
@@ -10,6 +11,7 @@ import dev.inmo.tgbotapi.types.InlineQueries.query.*
 internal suspend inline fun <reified T : InlineQuery> BehaviourContext.onInlineQuery(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     noinline additionalFilter: (suspend (T) -> Boolean)? = null,
+    performInParallel: Boolean = true,
     noinline scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, T>
 ) = flowsUpdatesFilter.expectFlow(bot) {
     it.asInlineQueryUpdate() ?.data ?.let { query ->
@@ -19,36 +21,42 @@ internal suspend inline fun <reified T : InlineQuery> BehaviourContext.onInlineQ
             null
         }
     }.let(::listOfNotNull)
-}.subscribeSafelyWithoutExceptions(scope) { triggerQuery ->
-    doInSubContextWithUpdatesFilter(
-        updatesFilter = if (includeFilterByChatInBehaviourSubContext) {
-            { it.sourceChat() ?.id ?.chatId == triggerQuery.from.id.chatId }
-        } else {
-            null
-        },
-        stopOnCompletion = false
-    ) {
-        scenarioReceiver(triggerQuery)
+}.subscribeSafelyWithoutExceptions(
+    scope,
+    optionallyWrapWithLaunch(performInParallel) { triggerQuery ->
+        doInSubContextWithUpdatesFilter(
+            updatesFilter = if (includeFilterByChatInBehaviourSubContext) {
+                { it.sourceChat() ?.id ?.chatId == triggerQuery.from.id.chatId }
+            } else {
+                null
+            },
+            stopOnCompletion = false
+        ) {
+            scenarioReceiver(triggerQuery)
+        }
     }
-}
+)
 
 
 suspend fun BehaviourContext.onAnyInlineQuery(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     additionalFilter: (suspend (InlineQuery) -> Boolean)? = null,
+    performInParallel: Boolean = true,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, InlineQuery>
-) = onInlineQuery(includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+) = onInlineQuery(includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)
 
 
 suspend fun BehaviourContext.onBaseInlineQuery(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     additionalFilter: (suspend (BaseInlineQuery) -> Boolean)? = null,
+    performInParallel: Boolean = true,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, BaseInlineQuery>
-) = onInlineQuery(includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+) = onInlineQuery(includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)
 
 
 suspend fun BehaviourContext.onLocationInlineQuery(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     additionalFilter: (suspend (LocationInlineQuery) -> Boolean)? = null,
+    performInParallel: Boolean = true,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, LocationInlineQuery>
-) = onInlineQuery(includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+) = onInlineQuery(includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)