+++ tgbotapi.extensions.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/triggers_handling/MediaGroupTriggers.kt
@@ -5,6 +5,7 @@ package dev.inmo.tgbotapi.extensions.behaviour_builder.triggers_handling
 import dev.inmo.micro_utils.coroutines.subscribeSafelyWithoutExceptions
 import dev.inmo.tgbotapi.extensions.behaviour_builder.*
 import dev.inmo.tgbotapi.extensions.behaviour_builder.expectations.expectFlow
+import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.optionallyWrapWithLaunch
 import dev.inmo.tgbotapi.extensions.utils.asSentMediaGroupUpdate
 import dev.inmo.tgbotapi.extensions.utils.extensions.sourceChat
 import dev.inmo.tgbotapi.extensions.utils.shortcuts.chat
@@ -18,6 +19,7 @@ import dev.inmo.tgbotapi.utils.PreviewFeature
 internal suspend inline fun <reified T : MediaGroupContent> BehaviourContext.buildMediaGroupTrigger(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     noinline additionalFilter: (suspend (List<MediaGroupMessage<T>>) -> Boolean)? = null,
+    performInParallel: Boolean = true,
     noinline scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, List<MediaGroupMessage<T>>>
 ) = flowsUpdatesFilter.expectFlow(bot) { update ->
     update.asSentMediaGroupUpdate() ?.data ?.let { mediaGroup ->
@@ -27,51 +29,61 @@ internal suspend inline fun <reified T : MediaGroupContent> BehaviourContext.bui
             null
         }
     } ?: emptyList()
-}.subscribeSafelyWithoutExceptions(scope) { mediaGroup ->
-    val mediaGroupChat = mediaGroup.chat!!
-    doInSubContextWithUpdatesFilter(
-        updatesFilter = if (includeFilterByChatInBehaviourSubContext) {
-            { it.sourceChat() ?.id ?.chatId == mediaGroupChat.id.chatId }
-        } else null,
-        stopOnCompletion = false
-    ) {
-        scenarioReceiver(mediaGroup)
+}.subscribeSafelyWithoutExceptions(
+    scope,
+    optionallyWrapWithLaunch(performInParallel) { mediaGroup ->
+        val mediaGroupChat = mediaGroup.chat!!
+        doInSubContextWithUpdatesFilter(
+            updatesFilter = if (includeFilterByChatInBehaviourSubContext) {
+                { it.sourceChat() ?.id ?.chatId == mediaGroupChat.id.chatId }
+            } else null,
+            stopOnCompletion = false
+        ) {
+            scenarioReceiver(mediaGroup)
+        }
     }
-}
+)
 
 suspend fun BehaviourContext.onMediaGroup(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     additionalFilter: (suspend (List<MediaGroupMessage<MediaGroupContent>>) -> Boolean)? = null,
+    performInParallel: Boolean = true,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, List<MediaGroupMessage<MediaGroupContent>>>
-) = buildMediaGroupTrigger(includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+) = buildMediaGroupTrigger(includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)
 suspend fun BehaviourContext.onPlaylist(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     additionalFilter: (suspend (List<MediaGroupMessage<AudioMediaGroupContent>>) -> Boolean)? = null,
+    performInParallel: Boolean = true,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, List<MediaGroupMessage<AudioMediaGroupContent>>>
-) = buildMediaGroupTrigger(includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+) = buildMediaGroupTrigger(includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)
 suspend fun BehaviourContext.onDocumentsGroup(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     additionalFilter: (suspend (List<MediaGroupMessage<DocumentMediaGroupContent>>) -> Boolean)? = null,
+    performInParallel: Boolean = true,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, List<MediaGroupMessage<DocumentMediaGroupContent>>>
-) = buildMediaGroupTrigger(includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+) = buildMediaGroupTrigger(includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)
 suspend fun BehaviourContext.onVisualGallery(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     additionalFilter: (suspend (List<MediaGroupMessage<VisualMediaGroupContent>>) -> Boolean)? = null,
+    performInParallel: Boolean = true,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, List<MediaGroupMessage<VisualMediaGroupContent>>>
-) = buildMediaGroupTrigger(includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+) = buildMediaGroupTrigger(includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)
 suspend fun BehaviourContext.onVisualMediaGroup(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     additionalFilter: (suspend (List<MediaGroupMessage<VisualMediaGroupContent>>) -> Boolean)? = null,
+    performInParallel: Boolean = true,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, List<MediaGroupMessage<VisualMediaGroupContent>>>
-) = onVisualGallery(includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+) = onVisualGallery(includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)
 suspend fun BehaviourContext.onPhotoGallery(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     additionalFilter: (suspend (List<MediaGroupMessage<PhotoContent>>) -> Boolean)? = null,
+    performInParallel: Boolean = true,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, List<MediaGroupMessage<PhotoContent>>>
-) = buildMediaGroupTrigger(includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+) = buildMediaGroupTrigger(includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)
 suspend fun BehaviourContext.onVideoGallery(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
     additionalFilter: (suspend (List<MediaGroupMessage<VideoContent>>) -> Boolean)? = null,
+    performInParallel: Boolean = true,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, List<MediaGroupMessage<VideoContent>>>
-) = buildMediaGroupTrigger(includeFilterByChatInBehaviourSubContext, additionalFilter, scenarioReceiver)
+) = buildMediaGroupTrigger(includeFilterByChatInBehaviourSubContext, additionalFilter, performInParallel, scenarioReceiver)